requirements.txt
appium-python-client==3.1.0
selenium==4.14.0
Pillow==10.1.0
pytesseract==0.3.10
adb-pywrapper==1.0.1

adb_pywrapper~=1.0.1
geopy~=2.4.1
setuptools~=68.2.0
gpxpy~=1.6.2